bluetooth: controller: Moved vendor specific HCI to HAL
Moved vendor HCI code to new HAL folder to allow out-of-tree
vendor specific HCI implementations.
